requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-distill==3.2.4
django-quill==0.1.0
django-quill-editor==0.1.40
django-rest-framework==0.1.0
djangorestframework==3.15.1
idna==3.7
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.2
